immediately return images from memory cache
there is always a short delay when fetching images, even when the image is in the memory cache. this is due to the image being fetched using an operation (so jumping back and forth between threads).

this commit keeps the existing operation intact (allowing it to handle the complex logic such as HTTP, disk cache, etc), but fetches the image from the memory cache in the UIImageView extension.

note: the memory cache is hit twice